Apply @jolelievre's comments
